requirements.txt
pandas
numpy
matplotlib
tqdm
openai
python-dotenv
langchain==0.1.17
langchain-community
langchain-openai